スクロール修正

+++ GFF2-4B/GFF2-4B/Player.h
@@ -111,7 +111,7 @@ public:
 	void Move();
 
 	//プレイヤーの状態の取得
-	int GetPlayerState() { return player_state; }
+	int GetPlayerState() { return (int)player_state; }
 
 	//プレイヤーの状態の更新
 	void UpdatePlayerState();
